Exception handling delete

+++ src/main/java/com/tiagopaulin/curso/services/UserService.java
@@ -2,9 +2,12 @@ package com.tiagopaulin.curso.services;
 
 import com.tiagopaulin.curso.entities.User;
 import com.tiagopaulin.curso.repositories.UserRepository;
+import com.tiagopaulin.curso.services.exceptions.DataBaseException;
 import com.tiagopaulin.curso.services.exceptions.ResourceNotFoundException;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -38,7 +41,21 @@ public class UserService {
 
     public void delete(Long id) {
 
-        userRepository.deleteById(id);
+        if (!userRepository.existsById(id)) {
+
+            throw new ResourceNotFoundException(id);
+
+        }
+
+        try {
+
+            userRepository.deleteById(id);
+
+        } catch (DataIntegrityViolationException e) {
+
+            throw new DataBaseException(e.getMessage());
+
+        }
 
     }
